+++ projects/smart-table/src/lib/components/smart-table/smart-table.component.ts
@@ -408,7 +408,7 @@ export class SmartTableComponent implements OnInit, OnDestroy {
   public exportToExcel() {
     this.pageChanging = true;
     let dataQuery = this.activeDataQuery;
-    if (!this.moduleConfig.options || !this.moduleConfig.options.exportWithFitlers) {
+    if (!this.moduleConfig.options || !this.moduleConfig.options.exportWithFilters) {
       dataQuery = {
         ...this.activeDataQuery,
         filters: [],